Change int to long in Main, SalesManager

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
     public static void  main(String[] args) {
-        int [] sales = new int[] {10, 200, 3000, 40000};
+        long [] sales = new long[] {10, 200, 3000, 40000};
         SalesManager salesManager = new SalesManager(sales);
 
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
 
     }
